+++ packages/falcon-front-kit/src/Account/index.ts
@@ -1,4 +1,4 @@
 export * from './SignInFormProvider';
 export * from './SignOutProvider';
 export * from './SignUpFormProvider';
-export * from './PasswordResetFormProvider';
+export * from './ForgotPasswordFormProvider';
